requirements.txt
# Python packages
python_dotenv
pip-tools
tiktoken

# Langchain
langchain
langchain_community
langchain_chroma
langchain_experimental
langchain_docling
langchain_huggingface

# FastAPI
fastapi
uvicorn
gunicorn
pydantic
sqlmodel
python_multipart
pypdf

# Database
pymongo

# Hugging Face
transformers
sentence_transformers
huggingface-hub
